+++ src/wrapper/index.ts
@@ -2,35 +2,50 @@
 
 import axios from "axios";
 
-interface Log10Config {
-  url: string;
-  token: string;
-  orgId: string;
-  tags?: string[];
-}
+import { SDKOptions } from "..";
+import { SDKHooks } from "../hooks/hooks";
 
-class Log10 {
-  private config: Log10Config;
+class Log10Wrapper {
+  private readonly options$: SDKOptions & { hooks?: SDKHooks };
+  private readonly tags: string[] = [];
 
-  constructor(config: Log10Config) {
-    this.config = config;
+  constructor(options: SDKOptions = {}, tags?: string[]) {
+    const opt = options as unknown;
+    let hooks: SDKHooks;
+    if (
+      typeof opt === "object" &&
+      opt != null &&
+      "hooks" in opt &&
+      opt.hooks instanceof SDKHooks
+    ) {
+      hooks = opt.hooks;
+    } else {
+      hooks = new SDKHooks();
+    }
+
+    this.options$ = { ...options, hooks };
+    this.tags = tags || [];
+    void this.options$;
   }
 
   async logCompletion(completion: any): Promise<void> {
     try {
       const response = await axios.post(
-        `${this.config.url}/api/v1/completions`,
+        `${this.options$.serverURL}/api/v1/completions`,
         {
           ...completion,
-          organization_id: this.config.orgId,
+          organization_id: this.options$.xLog10Organization,
           kind: "chat",
           status: "finished",
-          tags: this.config.tags || [],
+          tags: this.tags || [],
         },
         {
           headers: {
-            "x-log10-token": this.config.token,
-            "x-log10-organization": this.config.orgId,
+            "x-log10-token":
+              typeof this.options$.log10Token === "function"
+                ? await this.options$.log10Token()
+                : this.options$.log10Token,
+            "x-log10-organization": this.options$.xLog10Organization,
             "Content-Type": "application/json",
           },
         }
@@ -45,16 +60,19 @@ class Log10 {
   async createFeedbackTask(taskSchema: any, name: string): Promise<string> {
     try {
       const response = await axios.post(
-        `${this.config.url}/api/v1/feedback_task`,
+        `${this.options$.serverURL}/api/v1/feedback_task`,
         {
           json_schema: taskSchema,
           name: name,
-          organization_id: this.config.orgId,
+          organization_id: this.options$.xLog10Organization,
         },
         {
           headers: {
-            "x-log10-token": this.config.token,
-            "x-log10-organization": this.config.orgId,
+            "x-log10-token":
+              typeof this.options$.log10Token === "function"
+                ? await this.options$.log10Token()
+                : this.options$.log10Token,
+            "x-log10-organization": this.options$.xLog10Organization,
             "Content-Type": "application/json",
           },
         }
@@ -73,17 +91,20 @@ class Log10 {
   ): Promise<void> {
     try {
       await axios.post(
-        `${this.config.url}/api/v1/feedback`,
+        `${this.options$.serverURL}/api/v1/feedback`,
         {
           task_id: taskId,
           json_values: values,
           completion_tags_selector: completionTagsSelector,
-          organization_id: this.config.orgId,
+          organization_id: this.options$.xLog10Organization,
         },
         {
           headers: {
-            "x-log10-token": this.config.token,
-            "x-log10-organization": this.config.orgId,
+            "x-log10-token":
+              typeof this.options$.log10Token === "function"
+                ? await this.options$.log10Token()
+                : this.options$.log10Token,
+            "x-log10-organization": this.options$.xLog10Organization,
             "Content-Type": "application/json",
           },
         }
@@ -94,7 +115,7 @@ class Log10 {
     }
   }
 
-  wrapper(client: any) {
+  wrap(client: any) {
     const ref = client.chat.completions.create;
     client.chat.completions.create = async (...args: any) => {
       const response = await ref.call(client.chat.completions, ...args);
@@ -109,4 +130,4 @@ class Log10 {
   }
 }
 
-export { Log10, Log10Config };
+export { Log10Wrapper };